Fix visibility in use statements and improve earlier fix for function/const in use statements.
Visibility: Added logic to check for visibility keywords and handle them correctly.

Function/const: The previous fix would bow out as soon as one of these was encountered. That is incorrect as the keyword without an alias is still invalid.
The current fix will move on to the alias to evaluate that instead or error on the function/const keyword if no alias is found.

Also: Removed the extra code which was added in #126 from the test file as it's no longer needed.